Added Structure edge collision
Quite a few changes in adding this, not only is collision now stupidly long (which I will optimize later), but both floor and ladder have been reworked to have pointers to the ladders/floors that are the sides of their respective groups. In order to accommodate this these pointers must be set in the buildLevel method, which there are a couple of examples of. I'll refine this later.