Extreme Makeover: Sound Edition
Sound is now fully functional in game.  Created new AudioCenter class to
manage sounds, which level holds an instance of.  Modified level so it
sort of matches up with sound.  If we ever get the game to go to other
levels, then we can impliment roundClear(), which is a method of
AudioCenter